# openapi.yaml
openapi: 3.0.0
info:
  title: Openstreetmap admin openapi-0.0
  description: |-
    Recycle map admin
  version: 0.0.0
servers:
  - url: http://127.0.0.1:3102/api
tags:
  - name: markers
    description: Operations related to markers

paths:
  /markers:
    get:
      tags:
        - markers
      summary: Get suggested markers
      description: Get suggested markers
      operationId: getMarkers
      responses:
        '200':
          $ref: '#/components/responses/Markers'
  /markers/{markerId}:
    parameters:
      - in: path
        name: markerId
        required: true
        schema:
          type: integer
    patch:
      tags:
        - markers
      summary: Update an existing marker
      description: Update an existing marker
      operationId: updateMarker
      requestBody:
        $ref: '#/components/requestBodies/UpdateMarker'
      responses:
        '204':
          description: No Content
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - markers
      summary: Delete suggested marker
      description: Delete suggested marker
      operationId: deleteMarker
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'

components:
  schemas:
    BadRequestDetails:
      type: object
      properties:
        location:
          type: string
          example: 'body'
        msg:
          type: string
          example: 'Field is required'
        param:
          type: string
          example: 'param'
        value: {}
    BadRequestError:
      type: object
      properties:
        message:
          type: string
          example: 'Invalid value'
      additionalProperties:
        $ref: '#/components/schemas/BadRequestDetails'
    NotFoundError:
      type: object
      properties:
        message:
          type: string
          example: 'Marker not found'
    UpdateMarkerRequest:
      type: object
      properties:
        position:
          type: object
          properties:
            approvedValue:
              type: array
              items:
                type: number
            suggestedValue:
              type: array
              items:
                type: array
                items:
                  type: number
        wasteTypes:
          type: object
          properties:
            approvedValue:
              type: array
              items:
                $ref: '#/components/schemas/WasteTypes'
            suggestedValue:
              type: array
              items:
                type: array
                items:
                  $ref: '#/components/schemas/WasteTypes'
        address:
          type: object
          properties:
            approvedValue:
              type: array
              items:
                type: string
              example: []
            suggestedValue:
              type: array
              items:
                type: array
                items:
                  type: string
    MarkersResponse:
      required:
        - id
        - position
        - wasteTypes
        - address
        - date
      type: object
      properties:
        id:
          type: string
          example: '57e7012c-2b6f-4e8b-ab86-e9fb91d4754e'
        position:
          type: object
          properties:
            approvedValue:
              type: array
              items:
                type: number
            suggestedValue:
              type: array
              items:
                type: array
                items:
                  type: number
          example: {approvedValue: [10.4, 10.4], suggestedValue: [[10.4, 10.4]]}
        wasteTypes:
          type: object
          properties:
            approvedValue:
              type: array
              items:
                $ref: '#/components/schemas/WasteTypes'
            suggestedValue:
              type: array
              items:
                type: array
                items:
                  $ref: '#/components/schemas/WasteTypes'
          example: {approvedValue: ['packing', 'plastic'], suggestedValue: [['packing', 'plastic']]}
        address:
          type: object
          properties:
            approvedValue:
              type: array
              items:
                type: string
            suggestedValue:
              type: array
              items:
                type: array
                items:
                  type: string
          example: {approvedValue: 'street, 4', suggestedValue: ['street, 4', 'street, 5']}
        date:
          type: string
          format: date-time
          example: '2023-03-11T18:33:35.687+00:00'
    WasteTypes:
      type: string
      enum: ['packing', 'plastic', 'batteries', 'carton', 'clothes', 'paper', 'glass']
  requestBodies:
    UpdateMarker:
      description: Update an existing marker
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdateMarkerRequest'
  responses:
    Markers:
      description: List of markers
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/MarkersResponse'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFoundError'